examples/python/requirements.txt
pyserial==3.4
pynmea2==1.19.0